% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_survival.R
\name{rate2survival}
\alias{rate2survival}
\title{Convert death probabilities to survival function}
\usage{
rate2survival(rates, ages, from = "prob", init_age = NULL, years = NULL)
}
\arguments{
\item{rates}{vector, matrix or 3D array of mortality rates with age
(on the rows) and calendar year or cohort (on the columns) and simulation
number (3rd dimension)}

\item{ages}{vector of ages for `rates`}

\item{from}{character string representing the type of mortality rate to be converted
from. Takes the following values: "central" for central death rates, "prob"
for 1-year death probabilities, "force" for force of mortality.}

\item{init_age}{initial age for which the survival function is to be calculated at. If not provided,
the survival function will be calculated for the smallest age supplied in `ages`}

\item{years}{optional vector of years for `rates`. If not supplied, then the column names
of `rates` will be preserved}
}
\value{
associated survival function as a 3D array if `rates` is an array, or as a matrix otherwise
}
\description{
Converts mortality rates to the associated survival function.
}
\details{
The survival function has survival time (starting from 0) on the rows.
}
\examples{
# consider the Kannisto completion method on male mortality rates
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
fitted_ages <- 76:90

completed_rates <- complete_old_age(
AUS_male_rates, ages, old_ages, method = "kannisto", type = "central",
control = list(fitted_ages = fitted_ages))

# compute survival function of an individual aged 55
all_ages <- 0:130
surv_func <- rate2survival(completed_rates, ages = all_ages, from = 'central', init_age = 55)
}
