% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docStMoMoExt.R
\docType{package}
\name{StMoMoExt}
\alias{StMoMoExt}
\title{StMoMoExt.}
\description{
An extension of the 'StMoMo' R package to produce mortality rates
at older ages and risk-free survival probabilities. It also includes
functions to simulate life paths and output summary statistics.
}
